pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devopsmind"
version = "1.0.0"
description = "Gamified DevOps learning platform — practice Linux, Bash, Git, Ansible, Docker, K8s, Terraform & Observability challenges."
authors = [{ name = "InfraForge Labs", email = "devopsmind@infraforgelabs.com" }]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-hcl2",
    "pyyaml",
    "rich",
    "click",
    "colorama",
    "requests"
]

[project.scripts]
devopsmind = "devopsmind.__main__:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
include = ["devopsmind", "devopsmind.*"]

[tool.setuptools.package-data]
"devopsmind" = ["*.py", "challenges/**/*"]
